+++ src/Factorial.Api/Repository/IRepository.cs
@@ -2,7 +2,7 @@ namespace Factorial.Api.Repository
 {
     public interface IRepository
     {
-         int? Get(int n);
-         void Insert(int n, int result);
+         ulong? Get(int n);
+         void Insert(int n, ulong result);
     }
 }
